SQUASH: Add eina_thread_id to get Eina_ThreadId from Eina_Thread #117
The assumption that `Eina_Thread` is a `pthread_t` was already present
upstream, so now `Eina_ThreadId` is defined to be just `Eina_Thread`.
The assumption that `pthread_t` is an integral type is, however, **wrong
according to the standards**, but that might not cause problems on our
supported platforms. If we at some point aim to be posix-compliant,
we'll have to revisit eina_threads with doubled attention.